move _print_elapsed_time into dedicated submodule
